+++ ChangeLog.md
@@ -18,6 +18,7 @@ Eventum Issue Tracking System
 - Improvements to messageId generator, make it use RandomLib (Elan Ruusamäe)
 - IRC bot improvements (Elan Ruusamäe, #82)
 - Use var/ path for writable data (Elan Ruusamäe, #81)
+- Pass array of changed fields to Workflow::handleCustomFieldsUpdated (Bryan Alsdorf)
 
 2015-10-13, Version 3.0.3
 -------------------------